refactor: Moved ProgramDiscussion/Live Configuration Models to program's app
refactor: Moved ProgramDiscussion/Live Configuration Models to program's app

fix: Resolved pylint issues

fix: Resolved pylint issues

fix: Resolved pylint issues
