+++ openedx/core/djangoapps/discussions/admin.py
@@ -7,8 +7,7 @@ from simple_history.admin import SimpleHistoryAdmin
 
 from openedx.core.djangoapps.config_model_utils.admin import StackedConfigModelAdmin
 
-from .forms import ProgramDiscussionsConfigurationForm, ProgramLiveConfigurationForm
-from .models import DiscussionsConfiguration, ProgramDiscussionsConfiguration, ProgramLiveConfiguration
+from .models import DiscussionsConfiguration
 from .models import ProviderFilter
 
 
@@ -19,30 +18,6 @@ class DiscussionsConfigurationAdmin(SimpleHistoryAdmin):
 
     search_fields = (
         'context_key',
-        'enabled',
-        'provider_type',
-    )
-    list_filter = (
-        'enabled',
-        'provider_type',
-    )
-
-
-class ProgramDiscussionsConfigurationAdmin(SimpleHistoryAdmin):
-    """
-    Customize the admin interface for the program discussions configuration
-    """
-    form = ProgramDiscussionsConfigurationForm
-
-    fieldsets = (
-        (None, {
-            'fields': ('program_uuid', 'enabled', 'lti_configuration', 'pii_share_username', 'pii_share_email',
-                       'provider_type'),
-        }),
-    )
-
-    search_fields = (
-        'program_uuid',
         'enabled',
         'provider_type',
     )
@@ -111,32 +86,5 @@ class ProviderFilterAdmin(StackedConfigModelAdmin):
         DenyListFilter,
     )
 
-
-class ProgramLiveConfigurationAdmin(SimpleHistoryAdmin):
-    """
-    Customize the admin interface for the program live configuration
-    """
-    form = ProgramLiveConfigurationForm
-
-    fieldsets = (
-        (None, {
-            'fields': ('program_uuid', 'enabled', 'lti_configuration', 'pii_share_username', 'pii_share_email',
-                       'provider_type'),
-        }),
-    )
-
-    search_fields = (
-        'program_uuid',
-        'enabled',
-        'provider_type',
-    )
-    list_filter = (
-        'enabled',
-        'provider_type',
-    )
-
-
 admin.site.register(DiscussionsConfiguration, DiscussionsConfigurationAdmin)
-admin.site.register(ProgramDiscussionsConfiguration, ProgramDiscussionsConfigurationAdmin)
-admin.site.register(ProgramLiveConfiguration, ProgramLiveConfigurationAdmin)
 admin.site.register(ProviderFilter, ProviderFilterAdmin)
